+++ test/test_industry.py
@@ -21,6 +21,10 @@ class TestTicker(unittest.TestCase):
         result = self.ticker.industry_ps_ratio()
         print(result)
 
+    def test_industry_pb_ratio(self):
+        result = self.ticker.industry_pb_ratio()
+        print(result)
+
     def test_ttm_net_income_common_stockholders(self):
         result = self.ticker.ttm_net_income_common_stockholders()
         print(result)