+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Change Log
 0.0.25
 -------
 - Support Industry P/S Ratio[`#60 <https://github.com/defeat-beta/defeatbeta-api/issues/60>`_]
+- Support Industry P/B Ratio[`#61 <https://github.com/defeat-beta/defeatbeta-api/issues/61>`_]
 
 0.0.24
 -------